+++ src/main/java/nextstep/subway/line/application/LineService.java
@@ -46,9 +46,7 @@ public class LineService {
     @Transactional(readOnly = true)
     public List<LineResponse> findLines() {
         List<Line> persistLines = lineRepository.findAll();
-        return persistLines.stream()
-                .map(line -> LineResponse.from(line))
-                .collect(Collectors.toList());
+        return LineResponse.ofList(persistLines);
     }
 
     @Transactional(readOnly = true)
@@ -61,9 +59,7 @@ public class LineService {
     public LineResponse findLineResponseById(Long id) {
         Line persistLine = findLineById(id);
         Sections sections = persistLine.getSections();
-        List<StationResponse> stations = sections.getOrderedStations(persistLine).stream()
-                .map(it -> StationResponse.of(it))
-                .collect(Collectors.toList());
+        List<StationResponse> stations = StationResponse.ofList(sections.getOrderedStations(persistLine));
         return LineResponse.of(persistLine, stations);
     }
