+++ src/main/java/nextstep/subway/line/dto/LineResponse.java
@@ -44,6 +44,12 @@ public class LineResponse {
         );
     }
 
+    public static List<LineResponse> ofList(List<Line> lines) {
+        return lines.stream()
+                .map(LineResponse::from)
+                .collect(Collectors.toList());
+    }
+
     public Long getId() {
         return id;
     }
