+++ src/main/java/nextstep/subway/member/domain/Member.java
@@ -1,6 +1,6 @@
 package nextstep.subway.member.domain;
 
-import nextstep.subway.BaseEntity;
+import nextstep.subway.common.BaseEntity;
 import nextstep.subway.auth.application.AuthorizationException;
 import org.apache.commons.lang3.StringUtils;
